+++ src/index.ts
@@ -1,6 +1,5 @@
 import { ApolloQueryPipe } from './ApolloQueryPipe';
 import { SelectPipe } from './SelectPipe';
-import { Apollo, ApolloQuery } from './ApolloDecorator';
 import { Angular2Apollo, defaultApolloClient } from './Angular2Apollo';
 import { ApolloQueryObservable } from './ApolloQueryObservable';
 import { ApolloModule } from './ApolloModule';
@@ -10,9 +9,7 @@ export const APOLLO_PROVIDERS: any[] = [
 ];
 
 export {
-  Apollo,
   ApolloModule,
-  ApolloQuery,
   ApolloQueryObservable,
   ApolloQueryPipe,
   SelectPipe,