+++ tests/ApolloDecorator/index.ts
@@ -1,3 +1,0 @@
-import './common';
-import './queries';
-import './mutations';