+++ tests/index.ts
@@ -5,4 +5,3 @@ import 'reflect-metadata';
 // tests
 import './Angular2Apollo';
 import './ApolloQueryPipe';
-import './ApolloDecorator';